+++ fpy2/frontend/__init__.py
@@ -5,9 +5,5 @@ This module contains a parser, syntax checker, and type checking
 for the FPy language.
 """
 
-from . import fpyast as fpyast
-
 from .decorator import fpy
-from .formatter import Formatter
-
-fpyast.set_default_formatter(Formatter())
+from .syntax_check import SyntaxCheck
